pyproject.toml
[project]
name = "aim_mcp_demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.2",
    "langchain-mcp-adapters>=0.1.11",
    "langchain-openai>=1.0.1",
    "langgraph>=0.3.27",
    "mcp[cli]>=1.6.0",
    "numpy>=2.2.4",
    "openai>=1.72.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "tavily-python>=0.5.4",
]
